refresh setup

+++ main.sh
@@ -220,7 +220,9 @@ DEBUG=${DEBUG-}
 FORCE_REBUILD=${FORCE_REBUILD-}
 DRYRUN=${DRYRUN-}
 NOREFRESH=${NOREFRESH-}
-NBPARALLEL=${NBPARALLEL-4}
+NBPARALLEL=${NBPARALLEL-2}
+SKIP_TAGS_REBUILD=${SKIP_TAGS_REBUILD-}
+SKIP_TAGS_REFRESH=${SKIP_TAGS_REFRESH-${SKIP_TAGS_REBUILD}}
 SKIP_IMAGES_SCAN=${SKIP_IMAGES_SCAN-}
 SKIP_MINOR_ES="((elasticsearch):.*([0-5]\.?){3}(-32bit.*)?)"
 SKIP_MINOR_ES2="$SKIP_MINOR_ES|(elasticsearch:(5\.[0-4]\.)|(6\.8\.[0-8])|(6\.[0-7])|(7\.9\.[0-2])|(7\.[0-8]))"
@@ -241,8 +243,8 @@ SKIP_OS="$SKIP_OS|(centos:(centos)?5)"
 SKIP_OS="$SKIP_OS|(fedora.*(modular|21))"
 SKIP_OS="$SKIP_OS|(traefik:((camembert|cancoillotte|cantal|chevrotin|faisselle|livarot|maroilles|montdor|morbier|picodon|raclette|reblochon|roquefort|tetedemoine)(-alpine)?|rc.*|(v?([0-9]+\.[0-9]+\.).*$)))"
 SKIP_OS="$SKIP_OS|(minio.*(armhf|aarch))"
+SKIP_PHP="(php:(5.4|5.3|.*(RC|-rc-).*))"
 SKIP_OS="$SKIP_OS)"
-SKIP_PHP="(php:(.*(RC|-rc-).*))"
 SKIP_WINDOWS="(.*(nanoserver|windows))"
 SKIP_MISC="(-?(on.?build)|pgrouting.*old)|seafile-mc:(7.0.1|7.0.2|7.0.3|7.0.4|7.0.5|7.1.3)|(dejavu:(v.*|1\..\.?.?|2\..\..)|3\.[1-3]\..|3.0.0|.*alpha.*$)"
 SKIP_NODE="((node):.*alpine3\..?.?)"
@@ -265,7 +267,6 @@ corpusops/elasticsearch
 corpusops/pgrouting
 corpusops/postgis
 "
-
 find_top_node_() {
     img=library/node
     if [ ! -e $img ];then return;fi
@@ -284,6 +285,7 @@ find_top_node_() {
 find_top_node() { (set +e && find_top_node_ && set -e;); }
 NODE_TOP="$(echo $(find_top_node))"
 MAILU_VERSiON=1.7
+
 BATCHED_IMAGES="\
 corpusops/mariadb/10.1\
  corpusops/mariadb/10.5\
@@ -345,7 +347,9 @@ corpusops/postgres/9-alpine\
  corpusops/elasticsearch/1-alpine\
  corpusops/elasticsearch/2-alpine::23
 "
+SKIP_REFRESH_ANCESTORS=${SKIP_REFRESH_ANCESTORS-}
 
+declare -A duplicated_tags
 declare -A registry_tokens
 declare -A registry_services
 
@@ -478,6 +482,7 @@ get_image_changeset() {
     echo "$ret"
 }
 
+do_gen_image() { gen_image "$@"; }
 gen_image() {
     local image=$1 tag=$2
     local ldir="$TOPDIR/$image/$tag"
@@ -504,7 +509,7 @@ gen_image() {
         local df="$folder/Dockerfile.override"
         if [ -e "$df" ];then dockerfiles="$dockerfiles $df" && break;fi
     done
-    local parts="from args argspost helpers pre base post clean cleanpost labels labelspost"
+    local parts="from args argspost helpers pre base post clean cleanpost extra labels labelspost"
     for order in $parts;do
         for folder in . .. ../../..;do
             local df="$folder/Dockerfile.$order"
@@ -562,10 +567,12 @@ do_get_namespace_tag() {
             # ubuntu-bare / postgis
             if [ -e $i/version ];then version=$( cat $i/version );break;fi
         done
-        echo "$repo/$tag:$version"
+        echo "$repo/$tag:$version" \
+            | sed -re "s/(-?(server)?-(web-vault|postgresql|mysql)):/-server:\3-/g"
     done
 }
 
+do_get_image_tags() { get_image_tags "$@"; }
 get_image_tags() {
     local n=$1
     local results="" result=""
@@ -579,7 +586,7 @@ get_image_tags() {
     else
         has_more=0
     fi
-    if [ $has_more -eq 0 ];then
+    if [[ -z ${SKIP_TAGS_REFRESH} ]] && [ $has_more -eq 0 ];then
         while [ $has_more -eq 0 ];do
             i=$((i+1))
             result=$( curl "${u}?page=${i}" 2>/dev/null \
@@ -590,10 +597,12 @@ get_image_tags() {
         if [ ! -e "$TOPDIR/$n" ];then mkdir -p "$TOPDIR/$n";fi
         printf "$results\n" | sort -V > "$t.raw"
     fi
+    if [[ -z ${SKIP_TAGS_REBUILD} ]];then
     rm -f "$t"
-    ( for i in $(cat "$t.raw");do
+    ( for j in $(cat "$t.raw");do for i in $j;do
         if is_skipped "$n:$i";then debug "Skipped: $n:$i";else printf "$i\n";fi
-      done | awk '!seen[$0]++' | sort -V ) >> "$t"
+      done;done | awk '!seen[$0]++' | sort -V ) >> "$t"
+    fi
     set -e
     if [ -e "$t" ];then cat "$t";fi
 }
@@ -628,10 +637,21 @@ do_clean_tags() {
 #     refresh_images library/ubuntu: only refresh ubuntu images
 do_refresh_images() {
     local imagess="${@:-$default_images}"
+    cp -vf local/corpusops.bootstrap/bin/cops_pkgmgr_install.sh helpers/
+    if ! ( grep -q corpusops/docker-images .git/config );then
+    if [ ! -e local/docker-images ];then
+        git clone https://github.com/corpusops/docker-images local/docker-images
+    fi
+    ( cd local/docker-images && git fetch --all && git reset --hard origin/master \
+      && cp -rf helpers        rootfs packages ../..; )
+#      && cp -rf helpers Dock* rootfs packages ../..; )
+    fi
     while read images;do
         for image in $images;do
             if [[ -n $image ]];then
-                make_tags $image
+                if [[ -z "${SKIP_MAKE_TAGS-}" ]];then
+                    make_tags $image
+                fi
                 do_clean_tags $image
             fi
         done
@@ -881,6 +901,7 @@ load_batched_images() {
     local counter=0
     local default_batchsize=$1
     shift
+    local batched_images="$(echo $@ |xargs -n1)"
     for i in $@;do
         local imgs=${i//::*}
         local batchsize=$default_batchsize
@@ -893,7 +914,7 @@ load_batched_images() {
             local subimages=$(do_list_image $img)
             if [[ -z $subimages ]];then break;fi
             for j in $subimages;do
-                if ! ( is_in_images $j );then
+                if ! ( is_in_images $j ) && ( echo "$batched_images" | egrep -q "^$j$");then
                     local space=" "
                     if [ `expr $counter % $batchsize` = 0 ];then
                         space=""
@@ -964,7 +985,7 @@ do_usage() {
 
 do_main() {
     local args=${@:-usage}
-    local actions="make_tags|refresh_corpusops|refresh_images|build|gen_travis|gen_gh|gen|list_images|clean_tags|get_namespace_tag"
+    local actions="make_tags|refresh_corpusops|refresh_images|build|gen_travis|gen_gh|gen|list_images|clean_tags|get_namespace_tag|gen_image|get_image_tags"
     actions="@($actions)"
     action=${1-};
     if [[ -n "$@" ]];then shift;fi
